debt: server/lib/db.js: extract schema SQL to db-schema.js
Moves the SCHEMA_SQL + AGGREGATE_SQL arrays (~74 lines of DDL) to
a sibling module. db.js stays the query API; db-schema.js is the
pure DDL list consumed by initSchema.

db.js: 616 → 544 lines. npm test:unit 782/782 pass. knip clean.
